Update move component
