+++ apps/client/src/components/Policy/PrivacyPolicy/Policy1/index.tsx
