+++ apps/client/src/components/Policy/PrivacyPolicy/Policy10/index.tsx
