+++ apps/client/src/components/Policy/PrivacyPolicy/Policy10/style.ts
