+++ apps/client/src/components/Policy/PrivacyPolicy/Policy12/index.tsx
