+++ apps/client/src/components/Policy/PrivacyPolicy/Policy12/style.ts
