+++ apps/client/src/components/Policy/PrivacyPolicy/Policy2/index.tsx
