+++ apps/client/src/components/Policy/PrivacyPolicy/Policy3/style.ts
