+++ apps/client/src/components/Policy/PrivacyPolicy/Policy4/style.ts
