+++ apps/client/src/components/Policy/PrivacyPolicy/Policy5/index.tsx
