+++ apps/client/src/components/Policy/PrivacyPolicy/Policy5/style.ts
