+++ apps/client/src/components/Policy/PrivacyPolicy/Policy6/index.tsx
