+++ apps/client/src/components/Policy/PrivacyPolicy/Policy7/index.tsx
