+++ apps/client/src/components/Policy/PrivacyPolicy/Policy8/style.ts
