+++ apps/client/src/components/Policy/PrivacyPolicy/Policy9/index.tsx
