+++ apps/client/src/components/Policy/PrivacyPolicy/index.tsx
